Refactor Quantity Fields to BigDecimal for Enhanced Precision and Consistency
- Updated ProductDto, PurchaseItemDto, and related entity classes to change quantity and remaining quantity fields from Long to BigDecimal, improving precision in quantity handling.
- Modified database schema to reflect these changes, ensuring that quantity fields in the database are now of type numeric with appropriate precision.
- Adjusted service methods in ProductQuantityService and PurchaseService to accommodate the new BigDecimal types, enhancing calculation accuracy and validation logic.

These changes collectively enhance the inventory management system's ability to handle quantities with greater precision, ensuring more reliable data processing and storage.

+++ src/main/java/com/inventory/dto/ProductDto.java
@@ -24,7 +24,7 @@ public class ProductDto {
     private String status;
     private BigDecimal weight;
     private ProductMainType type;
-    private Long remainingQuantity;
+    private BigDecimal remainingQuantity;
     private Long clientId;
     private Integer page = 0;
     private Integer size = 10;
